requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pymysql==1.1.0
anthropic==0.18.0
pydantic-settings==2.1.0
python-dotenv==1.0.1